refactor(entry): serviceとstatisticsを責務分割 (#1308)
