+++ apps/product/src/features/entry/server/entry-service-error.ts
@@ -1,0 +1,8 @@
+import { ServiceError } from '@/lib/trpc/errors';
+
+export class EntryServiceError extends ServiceError {
+  constructor(code: string, message: string) {
+    super(code, message);
+    this.name = 'EntryServiceError';
+  }
+}
