+++ apps/product/src/features/entry/server/statistics.ts
@@ -1,810 +1,12 @@
-/**
- * Entries Statistics Router
- *
- * 統計・分析用のデータ集約エンドポイント
- */
-
-import * as Sentry from '@sentry/nextjs';
-import type { SupabaseClient } from '@supabase/supabase-js';
-import { TRPCError } from '@trpc/server';
-import { formatInTimeZone } from 'date-fns-tz';
-import { z } from 'zod';
-
-import { logger } from '@/lib/logger';
-import { traceDbQuery } from '@/lib/sentry/trace';
-import { createTRPCRouter, proProcedure, protectedProcedure } from '@/lib/trpc/procedures';
-import type { Database } from '@dayopt/database';
-
-import {
-  aggregateDayOfWeekDistribution,
-  aggregateHourlyDistribution,
-  aggregateMonthlyTrend,
-  aggregateTagStats,
-  calculateStreak,
-  type EstimationAccuracyDbRow,
-  getMonthlyStartDate,
-  transformEstimationAccuracy,
-} from '../domain';
-
-import { transformEnergyMapResponse } from './statistics-energy-map-transform';
-import {
-  unpackBlankRate,
-  unpackContextSwitches,
-  unpackCumulativeTime,
-  unpackEntryRate,
-} from './statistics-kpi-unpackers';
-import { transformStatsOverviewResponse } from './statistics-overview-transform';
-import { transformTimeByTagResponse } from './statistics-time-by-tag-transform';
-
-/**
- * ユーザーのタイムゾーンで「今日」の日付文字列（YYYY-MM-DD）を返す
- */
-function getTodayInTimezone(timezone: string): string {
-  return formatInTimeZone(new Date(), timezone, 'yyyy-MM-dd');
-}
-
-/**
- * supabase から user_settings.timezone を取得する（失敗時は 'UTC' にフォールバック）
- */
-async function getUserTimezone(
-  supabase: SupabaseClient<Database>,
-  userId: string,
-): Promise<string> {
-  const { data } = await supabase
-    .from('user_settings')
-    .select('timezone')
-    .eq('user_id', userId)
-    .single();
-  return (data?.timezone as string | null | undefined) ?? 'UTC';
-}
-
-// exactOptionalPropertyTypes 対応: undefined値を除外して optional params を安全に渡す
-type StripUndefinedValues<T> = { [K in keyof T]: Exclude<T[K], undefined> };
-function stripUndefined<T extends Record<string, unknown>>(obj: T): StripUndefinedValues<T> {
-  return Object.fromEntries(Object.entries(obj).filter(([, v]) => v !== undefined)) as never;
-}
-
-/** 統計クエリの共通エラーハンドラー */
-function handleStatsError(operation: string, error: unknown): never {
-  // TRPCError も Sentry に報告してから再スロー（内側throwのDB起因エラーを補足）
-  if (error instanceof TRPCError) {
-    Sentry.captureException(error.cause ?? error, {
-      tags: { source: 'statistics_router', operation },
-    });
-    throw error;
-  }
-
-  Sentry.captureException(error, {
-    tags: { source: 'statistics_router', operation },
-  });
-  logger.error(`Statistics ${operation} failed`, {
-    error: error instanceof Error ? error.message : String(error),
-  });
-
-  throw new TRPCError({
-    code: 'INTERNAL_SERVER_ERROR',
-    message: `Failed to fetch statistics (${operation}): ${error instanceof Error ? error.message : String(error)}`,
-    cause: error,
-  });
-}
-
-// =============================================================================
-// Schemas & Types
-// =============================================================================
-
-/** 期間フィルター用の共通入力スキーマ */
-const dateRangeInput = z.object({
-  startDate: z.string().datetime({ offset: true }).optional(),
-  endDate: z.string().datetime({ offset: true }).optional(),
-});
-
-// =============================================================================
-// Router
-// =============================================================================
-
-/** エントリ統計・分析用tRPCルーター（タグ統計・完了率・時間集計などを提供） */
-export const entriesStatisticsRouter = createTRPCRouter({
-  // ---------------------------------------------------------------------------
-  // General Statistics
-  // ---------------------------------------------------------------------------
-
-  /** Get tag statistics (entry count and last used date) */
-  getTagStats: protectedProcedure
-    .meta({ description: 'タグ別統計取得（エントリ数・最終使用日）' })
-    .query(async ({ ctx }) => {
-      try {
-        const { supabase, userId } = ctx;
-
-        const { data, error } = await traceDbQuery('stats.get_tag_stats', async () =>
-          supabase.rpc('get_tag_stats', { p_user_id: userId }),
-        );
-
-        if (error) {
-          throw new TRPCError({
-            code: 'INTERNAL_SERVER_ERROR',
-            message: `Failed to fetch tag statistics: ${error.message}`,
-            cause: error,
-          });
-        }
-
-        return aggregateTagStats(data);
-      } catch (error) {
-        handleStatsError('getTagStats', error);
-      }
-    }),
-
-  /** Get time spent per tag (DB function) */
-  getTimeByTag: protectedProcedure
-    .meta({ description: 'タグ別時間集計（期間フィルタ対応）' })
-    .input(dateRangeInput.optional())
-    .query(async ({ ctx, input }) => {
-      try {
-        const { supabase, userId } = ctx;
-
-        const { data, error } = await traceDbQuery('stats.get_time_by_tag', async () =>
-          supabase.rpc(
-            'get_time_by_tag',
-            stripUndefined({
-              p_user_id: userId,
-              p_start_date: input?.startDate,
-              p_end_date: input?.endDate,
-            }),
-          ),
-        );
-
-        if (error) {
-          throw new TRPCError({
-            code: 'INTERNAL_SERVER_ERROR',
-            message: `Failed to fetch time by tag: ${error.message}`,
-            cause: error,
-          });
-        }
-
-        return transformTimeByTagResponse(data);
-      } catch (error) {
-        handleStatsError('getTimeByTag', error);
-      }
-    }),
-
-  /** Get daily hours for heatmap (DB function) */
-  getDailyHours: protectedProcedure
-    .meta({ description: '日別記録時間取得（ヒートマップ用）' })
-    .input(z.object({ year: z.number().int().min(2000).max(2100) }))
-    .query(async ({ ctx, input }) => {
-      try {
-        const { supabase, userId } = ctx;
-        const { year } = input;
-
-        const { data, error } = await traceDbQuery('stats.get_daily_hours', async () =>
-          supabase.rpc('get_daily_hours', {
-            p_user_id: userId,
-            p_year: year,
-          }),
-        );
-
-        if (error) {
-          throw new TRPCError({
-            code: 'INTERNAL_SERVER_ERROR',
-            message: `Failed to fetch daily hours: ${error.message}`,
-            cause: error,
-          });
-        }
-
-        return data ?? [];
-      } catch (error) {
-        handleStatsError('getDailyHours', error);
-      }
-    }),
-
-  /** Get hourly distribution (DB function) */
-  getHourlyDistribution: proProcedure
-    .meta({ description: '時間帯別分布取得（2時間スロット）' })
-    .input(dateRangeInput.optional())
-    .query(async ({ ctx, input }) => {
-      try {
-        const { supabase, userId } = ctx;
-
-        const { data, error } = await traceDbQuery('stats.get_hourly_distribution', async () =>
-          supabase.rpc(
-            'get_hourly_distribution',
-            stripUndefined({
-              p_user_id: userId,
-              p_start_date: input?.startDate,
-              p_end_date: input?.endDate,
-            }),
-          ),
-        );
-
-        if (error) {
-          throw new TRPCError({
-            code: 'INTERNAL_SERVER_ERROR',
-            message: `Failed to fetch hourly distribution: ${error.message}`,
-            cause: error,
-          });
-        }
-
-        return aggregateHourlyDistribution(data ?? []);
-      } catch (error) {
-        handleStatsError('getHourlyDistribution', error);
-      }
-    }),
-
-  /** Get day of week distribution (DB function) */
-  getDayOfWeekDistribution: proProcedure
-    .meta({ description: '曜日別分布取得（月曜始まり）' })
-    .input(dateRangeInput.optional())
-    .query(async ({ ctx, input }) => {
-      try {
-        const { supabase, userId } = ctx;
-
-        const { data, error } = await traceDbQuery('stats.get_dow_distribution', async () =>
-          supabase.rpc(
-            'get_dow_distribution',
-            stripUndefined({
-              p_user_id: userId,
-              p_start_date: input?.startDate,
-              p_end_date: input?.endDate,
-            }),
-          ),
-        );
-
-        if (error) {
-          throw new TRPCError({
-            code: 'INTERNAL_SERVER_ERROR',
-            message: `Failed to fetch day-of-week distribution: ${error.message}`,
-            cause: error,
-          });
-        }
-
-        return aggregateDayOfWeekDistribution(data ?? []);
-      } catch (error) {
-        handleStatsError('getDayOfWeekDistribution', error);
-      }
-    }),
-
-  /** Get monthly trend (DB function) */
-  getMonthlyTrend: proProcedure
-    .meta({ description: '月別トレンド取得（デフォルト12ヶ月）' })
-    .input(z.object({ months: z.number().min(1).max(120).optional() }).optional())
-    .query(async ({ ctx, input }) => {
-      try {
-        const { supabase, userId } = ctx;
-        const monthCount = input?.months ?? 12;
-
-        // ユーザーのタイムゾーンで「現在の年月」を決定する
-        const timezone = await getUserTimezone(supabase, userId);
-        const nowStr = formatInTimeZone(new Date(), timezone, 'yyyy-MM');
-        const [nowYear, nowMonth] = nowStr.split('-').map(Number) as [number, number];
-
-        const startDate = getMonthlyStartDate(nowYear, nowMonth, monthCount);
-
-        const { data, error } = await traceDbQuery('stats.get_monthly_hours', async () =>
-          supabase.rpc('get_monthly_hours', {
-            p_user_id: userId,
-            p_start_date: startDate.toISOString(),
-          }),
-        );
-
-        if (error) {
-          throw new TRPCError({
-            code: 'INTERNAL_SERVER_ERROR',
-            message: `Failed to fetch monthly trend: ${error.message}`,
-            cause: error,
-          });
-        }
-
-        return aggregateMonthlyTrend(data ?? [], nowYear, nowMonth, monthCount);
-      } catch (error) {
-        handleStatsError('getMonthlyTrend', error);
-      }
-    }),
-
-  // ---------------------------------------------------------------------------
-  // Phase 1: KPI Metrics
-  // ---------------------------------------------------------------------------
-
-  /** エントリ率: origin='planned' / 全エントリ */
-  getEntryRate: proProcedure
-    .meta({ description: 'エントリ率KPI（計画エントリ / 全エントリ）' })
-    .input(dateRangeInput)
-    .query(async ({ ctx, input }) => {
-      try {
-        const { supabase, userId } = ctx;
-
-        const { data, error } = await traceDbQuery('stats.get_plan_rate', async () =>
-          supabase.rpc(
-            'get_plan_rate',
-            stripUndefined({
-              p_user_id: userId,
-              p_start_date: input.startDate,
-              p_end_date: input.endDate,
-            }),
-          ),
-        );
-
-        if (error) {
-          throw new TRPCError({
-            code: 'INTERNAL_SERVER_ERROR',
-            message: `Failed to fetch planning rate: ${error.message}`,
-            cause: error,
-          });
-        }
-
-        return unpackEntryRate(data);
-      } catch (error) {
-        handleStatsError('getEntryRate', error);
-      }
-    }),
-
-  /** 見積もり精度: タグ別の予定時間 vs 実績時間 */
-  getEstimationAccuracy: proProcedure
-    .meta({ description: '見積もり精度KPI（タグ別の予定vs実績）' })
-    .input(dateRangeInput)
-    .query(async ({ ctx, input }) => {
-      try {
-        const { supabase, userId } = ctx;
-
-        const { data, error } = await traceDbQuery('stats.get_estimation_accuracy', async () =>
-          supabase.rpc(
-            'get_estimation_accuracy',
-            stripUndefined({
-              p_user_id: userId,
-              p_start_date: input.startDate,
-              p_end_date: input.endDate,
-            }),
-          ),
-        );
-
-        if (error) {
-          throw new TRPCError({
-            code: 'INTERNAL_SERVER_ERROR',
-            message: `Failed to fetch estimation accuracy: ${error.message}`,
-            cause: error,
-          });
-        }
-
-        return transformEstimationAccuracy((data ?? []) as ReadonlyArray<EstimationAccuracyDbRow>);
-      } catch (error) {
-        handleStatsError('getEstimationAccuracy', error);
-      }
-    }),
-
-  /** エネルギーマップ: 時間帯×曜日の活動分布（既存DB関数のラッパー） */
-  getEnergyMap: proProcedure
-    .meta({ description: 'エネルギーマップ（時間帯×曜日の活動分布）' })
-    .input(dateRangeInput)
-    .query(async ({ ctx, input }) => {
-      try {
-        const { supabase, userId } = ctx;
-
-        const { data, error } = await traceDbQuery('stats.get_energy_map', async () =>
-          supabase.rpc('get_energy_map', {
-            p_user_id: userId,
-            p_start: input.startDate ?? '2000-01-01',
-            p_end: input.endDate ?? '2099-12-31',
-          }),
-        );
-
-        if (error) {
-          throw new TRPCError({
-            code: 'INTERNAL_SERVER_ERROR',
-            message: `Failed to fetch energy map: ${error.message}`,
-            cause: error,
-          });
-        }
-
-        return transformEnergyMapResponse(data);
-      } catch (error) {
-        handleStatsError('getEnergyMap', error);
-      }
-    }),
-
-  /** コンテキストスイッチ: 連続エントリ間のタグ変化回数 */
-  getContextSwitches: proProcedure
-    .meta({ description: 'コンテキストスイッチ回数（タグ変化頻度）' })
-    .input(dateRangeInput)
-    .query(async ({ ctx, input }) => {
-      try {
-        const { supabase, userId } = ctx;
-
-        const { data, error } = await traceDbQuery('stats.get_context_switches', async () =>
-          supabase.rpc(
-            'get_context_switches',
-            stripUndefined({
-              p_user_id: userId,
-              p_start_date: input.startDate,
-              p_end_date: input.endDate,
-            }),
-          ),
-        );
-
-        if (error) {
-          throw new TRPCError({
-            code: 'INTERNAL_SERVER_ERROR',
-            message: `Failed to fetch context switches: ${error.message}`,
-            cause: error,
-          });
-        }
-
-        return unpackContextSwitches(data);
-      } catch (error) {
-        handleStatsError('getContextSwitches', error);
-      }
-    }),
-
-  /** 空白率: 活動可能時間のうちスケジュールされていない時間の割合 */
-  getBlankRate: proProcedure
-    .meta({ description: '空白率KPI（未スケジュール時間の割合）' })
-    .input(
-      dateRangeInput.extend({
-        wakeHour: z.number().min(0).max(23).default(7),
-        sleepHour: z.number().min(0).max(23).default(23),
-      }),
-    )
-    .query(async ({ ctx, input }) => {
-      try {
-        const { supabase, userId } = ctx;
-
-        const { data, error } = await traceDbQuery('stats.get_blank_rate', async () =>
-          supabase.rpc(
-            'get_blank_rate',
-            stripUndefined({
-              p_user_id: userId,
-              p_start_date: input.startDate,
-              p_end_date: input.endDate,
-              p_wake_hour: input.wakeHour,
-              p_sleep_hour: input.sleepHour,
-            }),
-          ),
-        );
-
-        if (error) {
-          throw new TRPCError({
-            code: 'INTERNAL_SERVER_ERROR',
-            message: `Failed to fetch idle rate: ${error.message}`,
-            cause: error,
-          });
-        }
-
-        return unpackBlankRate(data);
-      } catch (error) {
-        handleStatsError('getBlankRate', error);
-      }
-    }),
-
-  /** 合計記録時間（分） */
-  getCumulativeTime: proProcedure
-    .meta({ description: '合計記録時間取得（分単位）' })
-    .input(dateRangeInput)
-    .query(async ({ ctx, input }) => {
-      try {
-        const { supabase, userId } = ctx;
-
-        const { data, error } = await traceDbQuery('stats.get_cumulative_time', async () =>
-          supabase.rpc(
-            'get_cumulative_time',
-            stripUndefined({
-              p_user_id: userId,
-              p_start_date: input.startDate,
-              p_end_date: input.endDate,
-            }),
-          ),
-        );
-
-        if (error) {
-          throw new TRPCError({
-            code: 'INTERNAL_SERVER_ERROR',
-            message: `Failed to fetch total recorded time: ${error.message}`,
-            cause: error,
-          });
-        }
-
-        return unpackCumulativeTime(data);
-      } catch (error) {
-        handleStatsError('getCumulativeTime', error);
-      }
-    }),
-
-  // ---------------------------------------------------------------------------
-  // Streak
-  // ---------------------------------------------------------------------------
-
-  /** 連続アクティブ日数（streak）を計算 */
-  getStreak: protectedProcedure
-    .meta({ description: '連続アクティブ日数（ストリーク）取得' })
-    .query(async ({ ctx }) => {
-      try {
-        const { supabase, userId } = ctx;
-
-        // ユーザーのタイムゾーンで「今日」を決定する
-        const timezone = await getUserTimezone(supabase, userId);
-        const todayStr = getTodayInTimezone(timezone);
-
-        // 過去365日分のアクティブ日を取得
-        const since = new Date(Date.now() - 365 * 24 * 60 * 60 * 1000);
-
-        const { data, error } = await traceDbQuery('stats.get_active_dates', async () =>
-          supabase.rpc('get_active_dates', {
-            p_user_id: userId,
-            p_start_date: formatInTimeZone(since, timezone, 'yyyy-MM-dd'),
-          }),
-        );
-
-        if (error) {
-          throw new TRPCError({
-            code: 'INTERNAL_SERVER_ERROR',
-            message: `Failed to fetch active days: ${error.message}`,
-            cause: error,
-          });
-        }
-
-        const streak = calculateStreak({
-          activeDates: data ?? [],
-          todayStr,
-          timezone,
-        });
-
-        return { streak };
-      } catch (error) {
-        handleStatsError('getStreak', error);
-      }
-    }),
-
-  // ---------------------------------------------------------------------------
-  // Unified KPI Summary (7 RPCs → 1 round-trip)
-  // ---------------------------------------------------------------------------
-
-  /** 全KPIを1クエリで取得 */
-  getStatsOverview: proProcedure
-    .meta({ description: '全KPIサマリー一括取得（7指標を1クエリ）' })
-    .input(
-      dateRangeInput.extend({
-        wakeHour: z.number().min(0).max(23).default(7),
-        sleepHour: z.number().min(0).max(23).default(23),
-      }),
-    )
-    .query(async ({ ctx, input }) => {
-      try {
-        const { supabase, userId } = ctx;
-
-        const { data, error } = await traceDbQuery('stats.get_kpi_summary', async () =>
-          supabase.rpc(
-            'get_stats_kpi_summary',
-            stripUndefined({
-              p_user_id: userId,
-              p_start_date: input.startDate,
-              p_end_date: input.endDate,
-              p_wake_hour: input.wakeHour,
-              p_sleep_hour: input.sleepHour,
-            }),
-          ),
-        );
-
-        if (error) {
-          throw new TRPCError({
-            code: 'INTERNAL_SERVER_ERROR',
-            message: `Failed to fetch KPI summary: ${error.message}`,
-            cause: error,
-          });
-        }
-
-        return transformStatsOverviewResponse(data);
-      } catch (error) {
-        handleStatsError('getStatsOverview', error);
-      }
-    }),
-
-  // ---------------------------------------------------------------------------
-  // Time P/L Data
-  // ---------------------------------------------------------------------------
-
-  /** Time P/L 用のタグ別予実データ + 日次ポイント */
-  getTimePL: protectedProcedure
-    .meta({ description: 'Time P/L データ取得（タグ別予実 + 日次累積）' })
-    .input(
-      z.object({
-        startDate: z.string().datetime({ offset: true }),
-        endDate: z.string().datetime({ offset: true }),
-        prevStart: z.string().datetime({ offset: true }).optional(),
-        prevEnd: z.string().datetime({ offset: true }).optional(),
-        wakeHour: z.number().min(0).max(23).default(7),
-        sleepHour: z.number().min(0).max(23).default(23),
-      }),
-    )
-    .query(async ({ ctx, input }) => {
-      try {
-        const { supabase, userId } = ctx;
-
-        // NOTE: get_time_pl_data は新規追加のため database.types.ts にまだ含まれない。
-        // 型生成後にこの as never キャストを除去する。
-        const { data, error } = await traceDbQuery('stats.get_time_pl_data', async () =>
-          supabase.rpc(
-            'get_time_pl_data' as never,
-            stripUndefined({
-              p_user_id: userId,
-              p_start_date: input.startDate,
-              p_end_date: input.endDate,
-              p_prev_start: input.prevStart,
-              p_prev_end: input.prevEnd,
-              p_wake_hour: input.wakeHour,
-              p_sleep_hour: input.sleepHour,
-            }) as never,
-          ),
-        );
-
-        if (error) {
-          throw new TRPCError({
-            code: 'INTERNAL_SERVER_ERROR',
-            message: `Failed to fetch time PL data: ${error.message}`,
-            cause: error,
-          });
-        }
-
-        return data as unknown as TimePLResponse;
-      } catch (error) {
-        handleStatsError('getTimePL', error);
-      }
-    }),
-
-  // ---------------------------------------------------------------------------
-  // Unified Stats Page Data (12 RPCs → 1 round-trip)
-  // ---------------------------------------------------------------------------
-
-  /** Stats ページ全データを 1 RPC で取得 */
-  getStatsPageData: proProcedure
-    .meta({ description: 'Stats全データ一括取得（12クエリ統合）' })
-    .input(
-      z.object({
-        startDate: z.string().datetime({ offset: true }),
-        endDate: z.string().datetime({ offset: true }),
-        prevStart: z.string().datetime({ offset: true }),
-        prevEnd: z.string().datetime({ offset: true }),
-        year: z.number().int().min(2000).max(2100),
-        monthlyMonths: z.number().int().min(1).max(24).default(3),
-        wakeHour: z.number().min(0).max(23).default(7),
-        sleepHour: z.number().min(0).max(23).default(23),
-      }),
-    )
-    .query(async ({ ctx, input }) => {
-      try {
-        const { supabase, userId } = ctx;
-
-        const { data, error } = await traceDbQuery('stats.get_stats_page_data', async () =>
-          supabase.rpc('get_stats_page_data', {
-            p_user_id: userId,
-            p_start_date: input.startDate,
-            p_end_date: input.endDate,
-            p_prev_start: input.prevStart,
-            p_prev_end: input.prevEnd,
-            p_year: input.year,
-            p_monthly_months: input.monthlyMonths,
-            p_wake_hour: input.wakeHour,
-            p_sleep_hour: input.sleepHour,
-          }),
-        );
-
-        if (error) {
-          throw new TRPCError({
-            code: 'INTERNAL_SERVER_ERROR',
-            message: `Failed to fetch stats page data: ${error.message}`,
-            cause: error,
-          });
-        }
-
-        // DB関数が返すJSONをそのまま返す（型はクライアント側で定義）
-        return data as unknown as StatsPageData;
-      } catch (error) {
-        handleStatsError('getStatsPageData', error);
-      }
-    }),
-});
-
-// ---------------------------------------------------------------------------
-// Types for unified stats response
-// ---------------------------------------------------------------------------
-
-interface StatsPageData {
-  overview: {
-    totalMinutes: number;
-    entryCount: number;
-    totalEntries: number;
-    plannedEntries: number;
-    planRate: number;
-  };
-  prevOverview: {
-    totalMinutes: number;
-    entryCount: number;
-    totalEntries: number;
-    plannedEntries: number;
-    planRate: number;
-  };
-  contextSwitches: {
-    totalSwitches: number;
-    avgPerDay: number;
-  };
-  blankRate: {
-    availableMinutes: number;
-    scheduledMinutes: number;
-    blankRate: number;
-  };
-  timeByTag: Array<{
-    tagId: string;
-    name: string;
-    color: string;
-    hours: number;
-  }>;
-  hourly: Array<{
-    hour: number;
-    totalMinutes: number;
-  }>;
-  dow: Array<{
-    dow: number;
-    totalMinutes: number;
-  }>;
-  energyMap: Array<{
-    hour: number;
-    dow: number;
-    totalMinutes: number;
-    entryCount: number;
-  }>;
-  estimationAccuracy: Array<{
-    tagId: string;
-    tagName: string;
-    tagColor: string;
-    avgPlannedMinutes: number;
-    avgActualMinutes: number;
-    avgDeviationMinutes: number;
-    entryCount: number;
-  }>;
-  prevEstimationAccuracy: Array<{
-    tagId: string;
-    tagName: string;
-    tagColor: string;
-    avgPlannedMinutes: number;
-    avgActualMinutes: number;
-    avgDeviationMinutes: number;
-    entryCount: number;
-  }>;
-  prevEnergyMap: Array<{
-    hour: number;
-    dow: number;
-    totalMinutes: number;
-    entryCount: number;
-  }>;
-  dailyHours: Array<{
-    day: string;
-    hours: number;
-  }>;
-  monthlyTrend: Array<{
-    month: string;
-    hours: number;
-  }>;
-}
-
-/** get_time_pl_data DB関数のレスポンス型 */
-interface TimePLResponse {
-  tags: Array<{
-    tagId: string;
-    tagName: string;
-    tagColor: string;
-    tagIcon: string | null;
-    budgetMinutes: number;
-    actualMinutes: number;
-    isPlanned: boolean;
-  }>;
-  prevTags: Array<{
-    tagId: string;
-    tagName: string;
-    tagColor: string;
-    tagIcon: string | null;
-    budgetMinutes: number;
-    actualMinutes: number;
-    isPlanned: boolean;
-  }>;
-  dailyPoints: Array<{
-    label: string;
-    budgetMinutes: number;
-    actualMinutes: number;
-  }>;
-  availableMinutes: number;
-}
+import { mergeRouters } from '@/lib/trpc/procedures';
+
+import { entriesStatisticsGeneralRouter } from './statistics-general-router';
+import { entriesStatisticsKpiRouter } from './statistics-kpi-router';
+import { entriesStatisticsSummaryRouter } from './statistics-summary-router';
+
+/** エントリ統計・分析用tRPCルーター */
+export const entriesStatisticsRouter = mergeRouters(
+  entriesStatisticsGeneralRouter,
+  entriesStatisticsKpiRouter,
+  entriesStatisticsSummaryRouter,
+);
